requirements.txt
transformers
datasets
python-dotenv
wandb
sentencepiece
protobuf==3.20
accelerate
matplotlib
tqdm
swifter